# q_branch/fine-grained-monitor/scenarios/payment-service/payment-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Payment processing service with fraud detection",
        "title": "payment-service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
        "/health": {
            "get": {
                "description": "Returns service health status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/internal/generate-report": {
            "post": {
                "description": "Generate a compliance report for the specified period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Generate compliance report",
                "parameters": [
                    {
                        "description": "Report Request",
                        "name": "report",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ReportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/pay": {
            "post": {
                "description": "Process a payment transaction with fraud detection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Process a payment",
                "parameters": [
                    {
                        "description": "Payment Request",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.PaymentRequest": {
            "type": "object",
            "required": [
                "amount",
                "card_holder",
                "card_number",
                "currency",
                "cvv",
                "expiry_date",
                "merchant_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "card_holder": {
                    "type": "string"
                },
                "card_number": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "cvv": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expiry_date": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
        "main.PaymentResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "fraud_score": {
                    "type": "number"
                },
                "merchant_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "processed_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
        "main.ReportRequest": {
            "type": "object",
            "required": [
                "end_date",
                "report_type",
                "start_date"
            ],
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                },
                "report_type": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "main.ReportResponse": {
            "type": "object",
            "properties": {
                "download_url": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "generated_at": {
                    "type": "string"
                },
                "record_count": {
                    "type": "integer"
                },
                "report_id": {
                    "type": "string"
                },
                "report_type": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        }
    }
}